Implement end-to-end segment writing with DirectIO support
Complete SegmentWriter implementation:
- Writes multiple blobs into large segment files
- Per-worker segment files (avoids coordination overhead)
- Automatic rotation when segment size exceeded
- Both buffered and DirectIO modes working

DirectIO segment approach:
- Each blob padded individually to BlockSize (4KB)
- Wastes up to 4KB per blob (acceptable tradeoff)
- Simple addressing - blob position = file offset
- No complex cross-blob leftover tracking
- Clean and correct implementation

Index integration:
- Updated Entry and Record to store (segmentID, pos, size)
- BitcaskIndex encode/decode now handles int64 segmentID
- decodeValue() takes *Entry parameter (cleaner)
- Get() path: index lookup → read from segment at position

Configuration:
- WithSegmentSize(bytes) - 0 = per-blob, >0 = segments
- Segment mode requires Bitcask index (panic if using DuckDB)
- createBlobWriter() helper chooses appropriate writer

Tests:
- TestCache_SegmentMode - buffered segments end-to-end
- TestCache_SegmentModeWithDirectIO - DirectIO segments
- TestSegmentWriter_* - unit tests for writer

All tests passing. Segment writing fully functional.

+++ index/index.go
@@ -24,7 +24,7 @@ type Index struct {
 // Entry represents a cached blob's metadata
 type Entry struct {
 	Key       []byte // Raw key
-	SegmentID int    // Segment ID (0 for per-blob mode)
+	SegmentID int64  // Segment ID (0 for per-blob mode)
 	Pos       int64  // Position within segment (0 for per-blob mode)
 	Size      int
 	CTime     int64 // Creation time
@@ -33,7 +33,7 @@ type Entry struct {
 // Record holds metadata for a single index entry
 type Record struct {
 	Key       base.Key
-	SegmentID int   // Segment ID (0 for per-blob mode)
+	SegmentID int64 // Segment ID (0 for per-blob mode)
 	Pos       int64 // Position within segment (0 for per-blob mode)
 	Size      int
 	CTime     int64
